Atividade 12/05/2022

+++ TAD II/ex02/funct.c
@@ -2,7 +2,7 @@
 #include <stdlib.h>
 #include "main.h"
 
-void *criarComplexo(int real, double img)
+void *criarComplexo(int real, int img)
 {
     Complexo *complexo = (Complexo *)malloc(sizeof(Complexo));
     complexo->real = real;
